feat: 이미지 page.tsx 생성
